pyproject.toml
[tool.poetry]
name = "mcp-server-planton"
version = "0.1.0"
description = "MCP server for Planton Cloud - Query environments, organizations, and cloud resources with user permissions"
authors = ["Planton Cloud <engineering@planton.cloud>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/plantoncloud-inc/mcp-server-planton"
repository = "https://github.com/plantoncloud-inc/mcp-server-planton"
documentation = "https://github.com/plantoncloud-inc/mcp-server-planton/tree/main/docs"
keywords = ["mcp", "model-context-protocol", "planton-cloud", "infrastructure"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
packages = [{include = "mcp_server_planton", from = "src"}]

[tool.poetry.scripts]
mcp-server-planton = "mcp_server_planton.server:main"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"

# MCP
mcp = "^1.3.2"

# gRPC
grpcio = "^1.68.1"
protobuf = "^6.32.0"

# Configuration
pydantic-settings = "^2.7.0"

# Planton Cloud API stubs from Buf registry
blintora-apis-protocolbuffers-python = { version = "*", source = "buf-build" }

[tool.poetry.group.dev.dependencies]
mypy = "*"
ruff = "*"

[[tool.poetry.source]]
name = "buf-build"
url = "https://buf.build/gen/python"
priority = "supplemental"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "blintora_apis_protocolbuffers_python.*"
ignore_errors = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

